requirements.txt
numpy==1.22.4
imutils==0.5.4
opencv-python==4.5.5.64
pillow==9.0.1